Добавление шапки на страницы с правилами и результатами
