+++ js/main.js
@@ -1,3 +1,3 @@
 import Application from './application';
 
-Application.showRules();
+Application.showIntro();
