+++ js/screens/rules-screen.js
@@ -1,44 +1,26 @@
-import AbstractView from '../abstract-view';
 import Application from './../application';
+import HeaderView from './../views/header-view';
+import RulesView from './../views/rules-view';
 
-class RulesView extends AbstractView {
+class RulesScreen {
   constructor() {
-    super();
+    this.header = new HeaderView(false);
+    this.content = new RulesView();
+
+    this.content.onClick = this.gameStart.bind(this);
+    this.element = document.createElement(`div`);
+    this.element.appendChild(this.header.element);
+    this.element.appendChild(this.content.element);
   }
 
   get template() {
-    return `
-      <section class="rules">
-        <h2 class="rules__title">Правила</h2>
-        <ul class="rules__description">
-          <li>Угадай 10 раз для каждого изображения фото
-            <img class="rules__icon" src="img/icon-photo.png" width="32" height="31" alt="Фото"> или рисунок
-            <img class="rules__icon" src="img/icon-paint.png" width="32" height="31" alt="Рисунок"></li>
-          <li>Фотографиями или рисунками могут быть оба изображения.</li>
-          <li>На каждую попытку отводится 30 секунд.</li>
-          <li>Ошибиться можно не более 3 раз.</li>
-        </ul>
-        <p class="rules__ready">Готовы?</p>
-        <form class="rules__form">
-          <input class="rules__input" type="text" placeholder="Ваше Имя">
-          <button class="rules__button  continue" type="submit" disabled>Go!</button>
-        </form>
-      </section>`;
+    return this.element;
   }
 
-  bind() {
-    const continueButton = this.element.querySelector(`.rules__button`);
-    const nameField = this.element.querySelector(`.rules__input`);
-
-    nameField.addEventListener(`input`, () => {
-      continueButton.disabled = nameField.value.length > 0 ? false : true;
-    });
-
-    continueButton.addEventListener(`click`, (evt) => {
-      evt.preventDefault();
-      Application.showGame(nameField.value);
-    });
+  gameStart(playerName) {
+    this.content.onClick = Application.showGame(playerName);
   }
+
 }
 
-export default RulesView;
+export default RulesScreen;
